Create Todo list tests
Creates two tests for the todo-list page.
The todo title test passes, but the second, display significant number of todos, does not pass.
The todo list page is created, along with the todo-list component.
The component path is added to the app routing module.
The todo-list component is given the proper title.
Co author: Alice K. Ellstrom ellst029.umn.edu