Add 'api/todos' test to todo.service.spec
Adds the test to ensure that the correct todos are called when getTodos() is called.
Adds the test to ensure that one request is made to the url.
Adds the tests to ensure that the request made to the url is a GET request.
All tests pass!
Refactors 'service' to todoService.
Co-author: Alice K. Ellstrom ellst029.umn.edu